+++ lib/minimizer.dart
@@ -1,8 +1,13 @@
-
 import 'minimizer_platform_interface.dart';
 
 class Minimizer {
-  Future<String?> getPlatformVersion() {
-    return MinimizerPlatform.instance.getPlatformVersion();
+  // Future<void> toPreviousApp() async {
+  //   await MinimizerPlatform.instance.toPreviousApp();
+  // }
+
+  static toPreviousApp() async {
+    try {
+      await MinimizerPlatform.instance.toPreviousApp();
+    } catch (_) {}
   }
 }